+++ pyinterpolate/kriging/utils/process.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from pyinterpolate.distance.distance import calc_point_to_point_distance
 from pyinterpolate.kriging.utils.kwarnings import ZerosMatrixWarning, LeastSquaresApproximationWarning
-from pyinterpolate.processing.select_values import select_kriging_data
+from pyinterpolate.processing.select_values import select_kriging_data, select_kriging_data_from_direction
 from pyinterpolate.variogram import TheoreticalVariogram
 
 
@@ -57,15 +57,21 @@ def get_predictions(theoretical_model: TheoreticalVariogram,
     if neighbors_range is None:
         neighbors_range = theoretical_model.rang
 
-    if theoretical_model.is_directional:
-        prepared_data = select_kriging_data_from_direction()
+    if theoretical_model.direction is not None:
+        prepared_data = select_kriging_data_from_direction(unknown_position=unknown_location,
+                                                           data_array=known_locations,
+                                                           neighbors_range=neighbors_range,
+                                                           direction=theoretical_model.direction,
+                                                           number_of_neighbors=no_neighbors,
+                                                           use_all_neighbors_in_range=use_all_neighbors_in_range)
     else:
         prepared_data = select_kriging_data(unknown_position=unknown_location,
                                             data_array=known_locations,
                                             neighbors_range=neighbors_range,
                                             number_of_neighbors=no_neighbors,
                                             use_all_neighbors_in_range=use_all_neighbors_in_range)
-        unknown_distances = prepared_data[:, -1]
+
+    unknown_distances = prepared_data[:, -1]
 
     n = len(prepared_data)
     k = theoretical_model.predict(unknown_distances)